# Add `unmeshed` flag to stat command